Replace old eval framework
